Until Concourse 3.10, `external-url` was needed to allow `fly execute` a task with local input to work.
This required hacks on our the Vagrant configuration and in general
made Concourse flacky
(see https://github.com/concourse/concourse/issues/2069).

Concourse 3.10 finally made this optional, so we do the same. More in
details: we don't use it at all in the salt formula Vagrantfile. We
leave the possibility, if pillar `concourse.web_fqdn` is set, to use it.